pyproject.toml
[project]
name = "neptune4-telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitaliy Pavlov",email = "v-2841@yandex.ru"}
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
